优化ios childViewController问题

+++ app/GogoTV_iOS/Source/GogoTV/UI/View/GBottomTabView.m
@@ -56,7 +56,7 @@
     
     [self setCheckWithIndex:0];
 }
--(void)setCheckWithIndex:(int)index{
+-(void)setCheckWithIndex:(NSInteger)index{
     [self clearAllItemCheck];
     GBottomTabItemView * item = [itemArray objectAtIndex:index];
     if (item) {
